feat: Rename EzFormProvider to EzForm
