+++ packages/easy-ui/src/index.ts
@@ -1,6 +1,6 @@
 export * from '@mui/material';
 
-export * from './EzFormProvider';
+export * from './EzForm';
 
 export * from './EzCheckbox';
 export * from './EzTextField';
